HTML + CSS
Did the markup and the styling for the calculator.

Note: if the calculator is resized to be less than 350x350px, it'll become a mess. Would be nice if I could figure out a way to scale the fonts without scripting. Plus, before even that, I need to make the margins, paddings and borders scaleabe, too!